+++ sdk/resourcemanager/network/armnetwork/models_serde.go
@@ -11151,8 +11151,8 @@ func (e *ExclusionManagedRuleSet) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
-// MarshalJSON implements the json.Marshaller interface for type ExplicitProxySettings.
-func (e ExplicitProxySettings) MarshalJSON() ([]byte, error) {
+// MarshalJSON implements the json.Marshaller interface for type ExplicitProxy.
+func (e ExplicitProxy) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]interface{})
 	populate(objectMap, "enableExplicitProxy", e.EnableExplicitProxy)
 	populate(objectMap, "enablePacFile", e.EnablePacFile)
@@ -11163,8 +11163,8 @@ func (e ExplicitProxySettings) MarshalJSON() ([]byte, error) {
 	return json.Marshal(objectMap)
 }
 
-// UnmarshalJSON implements the json.Unmarshaller interface for type ExplicitProxySettings.
-func (e *ExplicitProxySettings) UnmarshalJSON(data []byte) error {
+// UnmarshalJSON implements the json.Unmarshaller interface for type ExplicitProxy.
+func (e *ExplicitProxy) UnmarshalJSON(data []byte) error {
 	var rawMsg map[string]json.RawMessage
 	if err := json.Unmarshal(data, &rawMsg); err != nil {
 		return fmt.Errorf("unmarshalling type %T: %v", e, err)
@@ -12234,6 +12234,7 @@ func (e ExpressRouteConnectionProperties) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]interface{})
 	populate(objectMap, "authorizationKey", e.AuthorizationKey)
 	populate(objectMap, "enableInternetSecurity", e.EnableInternetSecurity)
+	populate(objectMap, "enablePrivateLinkFastPath", e.EnablePrivateLinkFastPath)
 	populate(objectMap, "expressRouteCircuitPeering", e.ExpressRouteCircuitPeering)
 	populate(objectMap, "expressRouteGatewayBypass", e.ExpressRouteGatewayBypass)
 	populate(objectMap, "provisioningState", e.ProvisioningState)
@@ -12256,6 +12257,9 @@ func (e *ExpressRouteConnectionProperties) UnmarshalJSON(data []byte) error {
 			delete(rawMsg, key)
 		case "enableInternetSecurity":
 			err = unpopulate(val, "EnableInternetSecurity", &e.EnableInternetSecurity)
+			delete(rawMsg, key)
+		case "enablePrivateLinkFastPath":
+			err = unpopulate(val, "EnablePrivateLinkFastPath", &e.EnablePrivateLinkFastPath)
 			delete(rawMsg, key)
 		case "expressRouteCircuitPeering":
 			err = unpopulate(val, "ExpressRouteCircuitPeering", &e.ExpressRouteCircuitPeering)
@@ -12936,6 +12940,7 @@ func (e *ExpressRouteLinkMacSecConfig) UnmarshalJSON(data []byte) error {
 func (e ExpressRouteLinkPropertiesFormat) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]interface{})
 	populate(objectMap, "adminState", e.AdminState)
+	populate(objectMap, "coloLocation", e.ColoLocation)
 	populate(objectMap, "connectorType", e.ConnectorType)
 	populate(objectMap, "interfaceName", e.InterfaceName)
 	populate(objectMap, "macSecConfig", e.MacSecConfig)
@@ -12957,6 +12962,9 @@ func (e *ExpressRouteLinkPropertiesFormat) UnmarshalJSON(data []byte) error {
 		switch key {
 		case "adminState":
 			err = unpopulate(val, "AdminState", &e.AdminState)
+			delete(rawMsg, key)
+		case "coloLocation":
+			err = unpopulate(val, "ColoLocation", &e.ColoLocation)
 			delete(rawMsg, key)
 		case "connectorType":
 			err = unpopulate(val, "ConnectorType", &e.ConnectorType)
@@ -13191,6 +13199,7 @@ func (e ExpressRoutePortPropertiesFormat) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]interface{})
 	populate(objectMap, "allocationDate", e.AllocationDate)
 	populate(objectMap, "bandwidthInGbps", e.BandwidthInGbps)
+	populate(objectMap, "billingType", e.BillingType)
 	populate(objectMap, "circuits", e.Circuits)
 	populate(objectMap, "encapsulation", e.Encapsulation)
 	populate(objectMap, "etherType", e.EtherType)
@@ -13217,6 +13226,9 @@ func (e *ExpressRoutePortPropertiesFormat) UnmarshalJSON(data []byte) error {
 			delete(rawMsg, key)
 		case "bandwidthInGbps":
 			err = unpopulate(val, "BandwidthInGbps", &e.BandwidthInGbps)
+			delete(rawMsg, key)
+		case "billingType":
+			err = unpopulate(val, "BillingType", &e.BillingType)
 			delete(rawMsg, key)
 		case "circuits":
 			err = unpopulate(val, "Circuits", &e.Circuits)
@@ -14256,7 +14268,7 @@ func (f FirewallPolicyPropertiesFormat) MarshalJSON() ([]byte, error) {
 	populate(objectMap, "basePolicy", f.BasePolicy)
 	populate(objectMap, "childPolicies", f.ChildPolicies)
 	populate(objectMap, "dnsSettings", f.DNSSettings)
-	populate(objectMap, "explicitProxySettings", f.ExplicitProxySettings)
+	populate(objectMap, "explicitProxy", f.ExplicitProxy)
 	populate(objectMap, "firewalls", f.Firewalls)
 	populate(objectMap, "insights", f.Insights)
 	populate(objectMap, "intrusionDetection", f.IntrusionDetection)
@@ -14289,8 +14301,8 @@ func (f *FirewallPolicyPropertiesFormat) UnmarshalJSON(data []byte) error {
 		case "dnsSettings":
 			err = unpopulate(val, "DNSSettings", &f.DNSSettings)
 			delete(rawMsg, key)
-		case "explicitProxySettings":
-			err = unpopulate(val, "ExplicitProxySettings", &f.ExplicitProxySettings)
+		case "explicitProxy":
+			err = unpopulate(val, "ExplicitProxy", &f.ExplicitProxy)
 			delete(rawMsg, key)
 		case "firewalls":
 			err = unpopulate(val, "Firewalls", &f.Firewalls)
@@ -17532,6 +17544,7 @@ func (i InterfacePropertiesFormat) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]interface{})
 	populate(objectMap, "auxiliaryMode", i.AuxiliaryMode)
 	populate(objectMap, "dnsSettings", i.DNSSettings)
+	populate(objectMap, "disableTcpStateTracking", i.DisableTCPStateTracking)
 	populate(objectMap, "dscpConfiguration", i.DscpConfiguration)
 	populate(objectMap, "enableAcceleratedNetworking", i.EnableAcceleratedNetworking)
 	populate(objectMap, "enableIPForwarding", i.EnableIPForwarding)
@@ -17567,6 +17580,9 @@ func (i *InterfacePropertiesFormat) UnmarshalJSON(data []byte) error {
 			delete(rawMsg, key)
 		case "dnsSettings":
 			err = unpopulate(val, "DNSSettings", &i.DNSSettings)
+			delete(rawMsg, key)
+		case "disableTcpStateTracking":
+			err = unpopulate(val, "DisableTCPStateTracking", &i.DisableTCPStateTracking)
 			delete(rawMsg, key)
 		case "dscpConfiguration":
 			err = unpopulate(val, "DscpConfiguration", &i.DscpConfiguration)
@@ -19111,6 +19127,7 @@ func (m *ManagedRuleGroupOverride) UnmarshalJSON(data []byte) error {
 // MarshalJSON implements the json.Marshaller interface for type ManagedRuleOverride.
 func (m ManagedRuleOverride) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]interface{})
+	populate(objectMap, "action", m.Action)
 	populate(objectMap, "ruleId", m.RuleID)
 	populate(objectMap, "state", m.State)
 	return json.Marshal(objectMap)
@@ -19125,6 +19142,9 @@ func (m *ManagedRuleOverride) UnmarshalJSON(data []byte) error {
 	for key, val := range rawMsg {
 		var err error
 		switch key {
+		case "action":
+			err = unpopulate(val, "Action", &m.Action)
+			delete(rawMsg, key)
 		case "ruleId":
 			err = unpopulate(val, "RuleID", &m.RuleID)
 			delete(rawMsg, key)
@@ -26764,6 +26784,99 @@ func (s *SubnetPropertiesFormat) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
+// MarshalJSON implements the json.Marshaller interface for type SwapResource.
+func (s SwapResource) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]interface{})
+	populate(objectMap, "id", s.ID)
+	populate(objectMap, "name", s.Name)
+	populate(objectMap, "properties", s.Properties)
+	populate(objectMap, "type", s.Type)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type SwapResource.
+func (s *SwapResource) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return fmt.Errorf("unmarshalling type %T: %v", s, err)
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "id":
+			err = unpopulate(val, "ID", &s.ID)
+			delete(rawMsg, key)
+		case "name":
+			err = unpopulate(val, "Name", &s.Name)
+			delete(rawMsg, key)
+		case "properties":
+			err = unpopulate(val, "Properties", &s.Properties)
+			delete(rawMsg, key)
+		case "type":
+			err = unpopulate(val, "Type", &s.Type)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return fmt.Errorf("unmarshalling type %T: %v", s, err)
+		}
+	}
+	return nil
+}
+
+// MarshalJSON implements the json.Marshaller interface for type SwapResourceListResult.
+func (s SwapResourceListResult) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]interface{})
+	populate(objectMap, "value", s.Value)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type SwapResourceListResult.
+func (s *SwapResourceListResult) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return fmt.Errorf("unmarshalling type %T: %v", s, err)
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "value":
+			err = unpopulate(val, "Value", &s.Value)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return fmt.Errorf("unmarshalling type %T: %v", s, err)
+		}
+	}
+	return nil
+}
+
+// MarshalJSON implements the json.Marshaller interface for type SwapResourceProperties.
+func (s SwapResourceProperties) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]interface{})
+	populate(objectMap, "slotType", s.SlotType)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type SwapResourceProperties.
+func (s *SwapResourceProperties) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return fmt.Errorf("unmarshalling type %T: %v", s, err)
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "slotType":
+			err = unpopulate(val, "SlotType", &s.SlotType)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return fmt.Errorf("unmarshalling type %T: %v", s, err)
+		}
+	}
+	return nil
+}
+
 // MarshalJSON implements the json.Marshaller interface for type SystemData.
 func (s SystemData) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]interface{})
@@ -27503,6 +27616,7 @@ func (v VPNClientConfiguration) MarshalJSON() ([]byte, error) {
 	populate(objectMap, "vpnClientProtocols", v.VPNClientProtocols)
 	populate(objectMap, "vpnClientRevokedCertificates", v.VPNClientRevokedCertificates)
 	populate(objectMap, "vpnClientRootCertificates", v.VPNClientRootCertificates)
+	populate(objectMap, "vngClientConnectionConfigurations", v.VngClientConnectionConfigurations)
 	return json.Marshal(objectMap)
 }
 
@@ -27550,6 +27664,9 @@ func (v *VPNClientConfiguration) UnmarshalJSON(data []byte) error {
 			delete(rawMsg, key)
 		case "vpnClientRootCertificates":
 			err = unpopulate(val, "VPNClientRootCertificates", &v.VPNClientRootCertificates)
+			delete(rawMsg, key)
+		case "vngClientConnectionConfigurations":
+			err = unpopulate(val, "VngClientConnectionConfigurations", &v.VngClientConnectionConfigurations)
 			delete(rawMsg, key)
 		}
 		if err != nil {
@@ -30722,6 +30839,7 @@ func (v VirtualNetworkGatewayConnectionListEntityPropertiesFormat) MarshalJSON()
 	populate(objectMap, "connectionType", v.ConnectionType)
 	populate(objectMap, "egressBytesTransferred", v.EgressBytesTransferred)
 	populate(objectMap, "enableBgp", v.EnableBgp)
+	populate(objectMap, "enablePrivateLinkFastPath", v.EnablePrivateLinkFastPath)
 	populate(objectMap, "expressRouteGatewayBypass", v.ExpressRouteGatewayBypass)
 	populate(objectMap, "gatewayCustomBgpIpAddresses", v.GatewayCustomBgpIPAddresses)
 	populate(objectMap, "ipsecPolicies", v.IPSecPolicies)
@@ -30769,6 +30887,9 @@ func (v *VirtualNetworkGatewayConnectionListEntityPropertiesFormat) UnmarshalJSO
 			delete(rawMsg, key)
 		case "enableBgp":
 			err = unpopulate(val, "EnableBgp", &v.EnableBgp)
+			delete(rawMsg, key)
+		case "enablePrivateLinkFastPath":
+			err = unpopulate(val, "EnablePrivateLinkFastPath", &v.EnablePrivateLinkFastPath)
 			delete(rawMsg, key)
 		case "expressRouteGatewayBypass":
 			err = unpopulate(val, "ExpressRouteGatewayBypass", &v.ExpressRouteGatewayBypass)
@@ -30866,6 +30987,7 @@ func (v VirtualNetworkGatewayConnectionPropertiesFormat) MarshalJSON() ([]byte, 
 	populate(objectMap, "egressBytesTransferred", v.EgressBytesTransferred)
 	populate(objectMap, "egressNatRules", v.EgressNatRules)
 	populate(objectMap, "enableBgp", v.EnableBgp)
+	populate(objectMap, "enablePrivateLinkFastPath", v.EnablePrivateLinkFastPath)
 	populate(objectMap, "expressRouteGatewayBypass", v.ExpressRouteGatewayBypass)
 	populate(objectMap, "gatewayCustomBgpIpAddresses", v.GatewayCustomBgpIPAddresses)
 	populate(objectMap, "ipsecPolicies", v.IPSecPolicies)
@@ -30921,6 +31043,9 @@ func (v *VirtualNetworkGatewayConnectionPropertiesFormat) UnmarshalJSON(data []b
 			delete(rawMsg, key)
 		case "enableBgp":
 			err = unpopulate(val, "EnableBgp", &v.EnableBgp)
+			delete(rawMsg, key)
+		case "enablePrivateLinkFastPath":
+			err = unpopulate(val, "EnablePrivateLinkFastPath", &v.EnablePrivateLinkFastPath)
 			delete(rawMsg, key)
 		case "expressRouteGatewayBypass":
 			err = unpopulate(val, "ExpressRouteGatewayBypass", &v.ExpressRouteGatewayBypass)
@@ -31215,6 +31340,123 @@ func (v *VirtualNetworkGatewayNatRuleProperties) UnmarshalJSON(data []byte) erro
 	return nil
 }
 
+// MarshalJSON implements the json.Marshaller interface for type VirtualNetworkGatewayPolicyGroup.
+func (v VirtualNetworkGatewayPolicyGroup) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]interface{})
+	populate(objectMap, "etag", v.Etag)
+	populate(objectMap, "id", v.ID)
+	populate(objectMap, "name", v.Name)
+	populate(objectMap, "properties", v.Properties)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type VirtualNetworkGatewayPolicyGroup.
+func (v *VirtualNetworkGatewayPolicyGroup) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return fmt.Errorf("unmarshalling type %T: %v", v, err)
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "etag":
+			err = unpopulate(val, "Etag", &v.Etag)
+			delete(rawMsg, key)
+		case "id":
+			err = unpopulate(val, "ID", &v.ID)
+			delete(rawMsg, key)
+		case "name":
+			err = unpopulate(val, "Name", &v.Name)
+			delete(rawMsg, key)
+		case "properties":
+			err = unpopulate(val, "Properties", &v.Properties)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return fmt.Errorf("unmarshalling type %T: %v", v, err)
+		}
+	}
+	return nil
+}
+
+// MarshalJSON implements the json.Marshaller interface for type VirtualNetworkGatewayPolicyGroupMember.
+func (v VirtualNetworkGatewayPolicyGroupMember) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]interface{})
+	populate(objectMap, "attributeType", v.AttributeType)
+	populate(objectMap, "attributeValue", v.AttributeValue)
+	populate(objectMap, "name", v.Name)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type VirtualNetworkGatewayPolicyGroupMember.
+func (v *VirtualNetworkGatewayPolicyGroupMember) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return fmt.Errorf("unmarshalling type %T: %v", v, err)
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "attributeType":
+			err = unpopulate(val, "AttributeType", &v.AttributeType)
+			delete(rawMsg, key)
+		case "attributeValue":
+			err = unpopulate(val, "AttributeValue", &v.AttributeValue)
+			delete(rawMsg, key)
+		case "name":
+			err = unpopulate(val, "Name", &v.Name)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return fmt.Errorf("unmarshalling type %T: %v", v, err)
+		}
+	}
+	return nil
+}
+
+// MarshalJSON implements the json.Marshaller interface for type VirtualNetworkGatewayPolicyGroupProperties.
+func (v VirtualNetworkGatewayPolicyGroupProperties) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]interface{})
+	populate(objectMap, "isDefault", v.IsDefault)
+	populate(objectMap, "policyMembers", v.PolicyMembers)
+	populate(objectMap, "priority", v.Priority)
+	populate(objectMap, "provisioningState", v.ProvisioningState)
+	populate(objectMap, "vngClientConnectionConfigurations", v.VngClientConnectionConfigurations)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type VirtualNetworkGatewayPolicyGroupProperties.
+func (v *VirtualNetworkGatewayPolicyGroupProperties) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return fmt.Errorf("unmarshalling type %T: %v", v, err)
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "isDefault":
+			err = unpopulate(val, "IsDefault", &v.IsDefault)
+			delete(rawMsg, key)
+		case "policyMembers":
+			err = unpopulate(val, "PolicyMembers", &v.PolicyMembers)
+			delete(rawMsg, key)
+		case "priority":
+			err = unpopulate(val, "Priority", &v.Priority)
+			delete(rawMsg, key)
+		case "provisioningState":
+			err = unpopulate(val, "ProvisioningState", &v.ProvisioningState)
+			delete(rawMsg, key)
+		case "vngClientConnectionConfigurations":
+			err = unpopulate(val, "VngClientConnectionConfigurations", &v.VngClientConnectionConfigurations)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return fmt.Errorf("unmarshalling type %T: %v", v, err)
+		}
+	}
+	return nil
+}
+
 // MarshalJSON implements the json.Marshaller interface for type VirtualNetworkGatewayPropertiesFormat.
 func (v VirtualNetworkGatewayPropertiesFormat) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]interface{})
@@ -31238,6 +31480,7 @@ func (v VirtualNetworkGatewayPropertiesFormat) MarshalJSON() ([]byte, error) {
 	populate(objectMap, "vpnClientConfiguration", v.VPNClientConfiguration)
 	populate(objectMap, "vpnGatewayGeneration", v.VPNGatewayGeneration)
 	populate(objectMap, "vpnType", v.VPNType)
+	populate(objectMap, "virtualNetworkGatewayPolicyGroups", v.VirtualNetworkGatewayPolicyGroups)
 	return json.Marshal(objectMap)
 }
 
@@ -31309,6 +31552,9 @@ func (v *VirtualNetworkGatewayPropertiesFormat) UnmarshalJSON(data []byte) error
 			delete(rawMsg, key)
 		case "vpnType":
 			err = unpopulate(val, "VPNType", &v.VPNType)
+			delete(rawMsg, key)
+		case "virtualNetworkGatewayPolicyGroups":
+			err = unpopulate(val, "VirtualNetworkGatewayPolicyGroups", &v.VirtualNetworkGatewayPolicyGroups)
 			delete(rawMsg, key)
 		}
 		if err != nil {
@@ -32332,6 +32578,80 @@ func (v *VnetRoute) UnmarshalJSON(data []byte) error {
 			delete(rawMsg, key)
 		case "staticRoutes":
 			err = unpopulate(val, "StaticRoutes", &v.StaticRoutes)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return fmt.Errorf("unmarshalling type %T: %v", v, err)
+		}
+	}
+	return nil
+}
+
+// MarshalJSON implements the json.Marshaller interface for type VngClientConnectionConfiguration.
+func (v VngClientConnectionConfiguration) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]interface{})
+	populate(objectMap, "etag", v.Etag)
+	populate(objectMap, "id", v.ID)
+	populate(objectMap, "name", v.Name)
+	populate(objectMap, "properties", v.Properties)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type VngClientConnectionConfiguration.
+func (v *VngClientConnectionConfiguration) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return fmt.Errorf("unmarshalling type %T: %v", v, err)
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "etag":
+			err = unpopulate(val, "Etag", &v.Etag)
+			delete(rawMsg, key)
+		case "id":
+			err = unpopulate(val, "ID", &v.ID)
+			delete(rawMsg, key)
+		case "name":
+			err = unpopulate(val, "Name", &v.Name)
+			delete(rawMsg, key)
+		case "properties":
+			err = unpopulate(val, "Properties", &v.Properties)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return fmt.Errorf("unmarshalling type %T: %v", v, err)
+		}
+	}
+	return nil
+}
+
+// MarshalJSON implements the json.Marshaller interface for type VngClientConnectionConfigurationProperties.
+func (v VngClientConnectionConfigurationProperties) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]interface{})
+	populate(objectMap, "provisioningState", v.ProvisioningState)
+	populate(objectMap, "vpnClientAddressPool", v.VPNClientAddressPool)
+	populate(objectMap, "virtualNetworkGatewayPolicyGroups", v.VirtualNetworkGatewayPolicyGroups)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type VngClientConnectionConfigurationProperties.
+func (v *VngClientConnectionConfigurationProperties) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return fmt.Errorf("unmarshalling type %T: %v", v, err)
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "provisioningState":
+			err = unpopulate(val, "ProvisioningState", &v.ProvisioningState)
+			delete(rawMsg, key)
+		case "vpnClientAddressPool":
+			err = unpopulate(val, "VPNClientAddressPool", &v.VPNClientAddressPool)
+			delete(rawMsg, key)
+		case "virtualNetworkGatewayPolicyGroups":
+			err = unpopulate(val, "VirtualNetworkGatewayPolicyGroups", &v.VirtualNetworkGatewayPolicyGroups)
 			delete(rawMsg, key)
 		}
 		if err != nil {