+++ sdk/resourcemanager/network/armnetwork/response_types.go
@@ -2416,6 +2416,21 @@ type VPNSitesConfigurationClientDownloadResponse struct {
 	// placeholder for future response values
 }
 
+// VipSwapClientCreateResponse contains the response from method VipSwapClient.Create.
+type VipSwapClientCreateResponse struct {
+	// placeholder for future response values
+}
+
+// VipSwapClientGetResponse contains the response from method VipSwapClient.Get.
+type VipSwapClientGetResponse struct {
+	SwapResource
+}
+
+// VipSwapClientListResponse contains the response from method VipSwapClient.List.
+type VipSwapClientListResponse struct {
+	SwapResourceListResult
+}
+
 // VirtualApplianceSKUsClientGetResponse contains the response from method VirtualApplianceSKUsClient.Get.
 type VirtualApplianceSKUsClientGetResponse struct {
 	VirtualApplianceSKU